api/requirements.txt
fastapi==0.110.0
GitPython==3.1.42
langchain==0.1.11
langchain_community==0.0.27
langchain_core==0.1.30
langchain_text_splitters==0.0.1
pydantic==2.6.2
qdrant_client==1.8.0
python-dotenv==1.0.1
uvicorn==0.14.0
sentence-transformers==2.4.0
openai==1.12.0
tiktoken==0.6.0
langchain_openai==0.0.8